refactor() Review loading core service into 4 main files, and created a database service
