+++ src/core/services/loading/batch_processor.py
